原材料需求计划 - 自测周维度生成原材料记录 Nick

+++ APS-Modules/aps-monthplan/src/main/java/com/zlt/aps/monthplan/raw/service/impl/RawMaterialRequirePlanServiceImpl.java
@@ -303,7 +303,7 @@ public class RawMaterialRequirePlanServiceImpl extends AbstractDocService<RawMat
 
         for (FactoryMonthPlanProdFinal plan : monthPlans) {
             // 获取总产量
-            Long totalQty = plan.getTotalQty();
+            Integer totalQty = plan.getTotalQty();
             if (totalQty == null || totalQty == 0) {
                 continue;
             }
